call to start game is now in add-player.js. created text boxes for player actions and player bets.

+++ junopoker/src/main/java/com/v1/junopoker/model/Player.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Player {
     private String username;
-    private int chipCount;
+    private float chipCount;
     private Card[] holeCards = null;
     private Hand hand = null;
     private boolean inHand = false;
